+++ src/status.rs
@@ -1,4 +1,5 @@
 use serde::Serialize;
+use std::collections::HashMap;
 use std::fmt;
 
 /// Message sent to the MQTT broker which later forwards it to Home Assistant
@@ -17,20 +18,17 @@ pub struct StatusMessage {
     /// Memory usage in %
     pub memory_usage: Option<f32>,
 
-    /// Net TX rate in KiB/s
-    pub net_tx: Option<f64>,
-
-    /// Net RX rate in KiB/s
-    pub net_rx: Option<f64>,
+    pub network: HashMap<String, NetworkStatus>,
 }
 
-impl StatusMessage {
-    pub fn off() -> StatusMessage {
-        StatusMessage {
-            available: "OFF",
-            ..Default::default()
-        }
-    }
+/// Network status
+#[derive(Serialize, Debug, Default)]
+pub struct NetworkStatus {
+    /// Net TX rate in KiB/s
+    pub tx: Option<f64>,
+
+    /// Net RX rate in KiB/s
+    pub rx: Option<f64>,
 }
 
 impl fmt::Display for StatusMessage {